64-bit improvements
git-svn-id: https://svn.code.sf.net/p/vidix/code/trunk@13 84fa5670-3129-0410-a1ff-e783189bc518

+++ vidix/libdha/libdha.c
@@ -70,10 +70,10 @@ void libdha_exit(const char *message, int level)
 #include "kernelhelper/dhahelper.h"
 
 static int mem=-1;
-void *map_phys_mem(unsigned base, unsigned size)
+void *map_phys_mem(unsigned long base, unsigned long size)
 {
 #ifdef ARCH_ALPHA
-/* TODO: move it in sysdep */
+/* TODO: move it into sysdep */
   base += bus_base();
 #endif
   if ( (mem = open("/dev/dhahelper",O_RDWR)) < 0)
@@ -102,10 +102,10 @@ void *map_phys_mem(unsigned base, unsigned size)
 #else
 
 static int mem=-1;
-void *map_phys_mem(unsigned base, unsigned size)
+void *map_phys_mem(unsigned long base, unsigned long size)
 {    
 #ifdef ARCH_ALPHA
-/* TODO: move it in sysdep */
+/* TODO: move it into sysdep */
   base += bus_base();
 #endif
   if ( (mem = open(DEV_MEM,O_RDWR)) == -1) {
@@ -115,7 +115,7 @@ void *map_phys_mem(unsigned base, unsigned size)
 }
 #endif /* CONFIG_DHAHELPER */
 
-void unmap_phys_mem(void *ptr, unsigned size)
+void unmap_phys_mem(void *ptr, unsigned long size)
 {
   int res=munmap(ptr,size) ;
   if (res == -1) { perror("libdha: munmap() failed") ; exit(1) ; }
